+++ flake.nix
@@ -69,6 +69,7 @@
               lockFile = ./Cargo.lock;
               outputHashes = {
                 "luna-0.5.0" = "sha256-YJsEKm7UykhqNQjhLPBgpmCHQZQhDLFVAENSD43SyAw=";
+                "rune-0.1.5" = "sha256-03yaB4bv+iXYGVStpGDCNdJQzYh/oiB8hRFrIKuMfvU=";
                 "tagdata-0.1.4" = "sha256-L/Fa0gS4hTrbaWvuvBB4NJ85xud8ixv7sSalLKRINQU=";
                 "vista-recall-0.1.2" = "sha256-Zi3pKV0pZNZh/XM8dE+uhHIJuieT977D0bAD91/qrhY=";
               };